requirements.txt
PyGithub>=2.1.1
click>=8.1.0
gitpython>=3.1.40
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
typing-extensions>=4.8.0
